requirements.txt
apturl==0.5.2
attrs==21.2.0
bcrypt==3.2.0
blinker==1.8.2
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debtcollector==3.0.0
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
eventlet==0.36.1
fasteners==0.14.1
filelock==3.13.4
Flask==3.0.0
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
future==0.18.2
greenlet==3.0.3
httplib2==0.20.2
idna==3.3
imageai==3.0.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.5
mininet==2.3.0
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.8
netaddr==1.2.1
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
oslo.config==9.4.0
oslo.i18n==6.3.0
ovs==3.1.2
paramiko==2.9.3
pbr==6.0.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
protobuf==3.12.4
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.6.1
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.7.3
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
repoze.lru==0.7
requests==2.25.1
rfc3986==2.0.0
Routes==2.5.1
rsa==4.9
ryu==4.34
SecretStorage==3.3.1
six==1.16.0
sortedcontainers==2.4.0
ssh-import-id==5.11
stevedore==5.2.0
systemd-python==234
Telethon==1.36.0
texttable==1.6.4
tinyrpc==1.1.7
ttystatus==0.38
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.25.1
wadllib==1.3.6
WebOb==1.8.7
websocket-client==1.2.3
Werkzeug==3.0.3
wrapt==1.16.0
xdg==5
xkit==0.0.0
zipp==1.0.0
